Ignore SyntaxError in non-literal settings
In #8 we used the Settings model to store the default locations in SS. There is
a function used in the administration tab that iterates over all the settings
and expects all the values to be Python literals which are evaluated with
`ast.literal_eval`.

This commits updates the function so it ignores SyntaxError exceptions which
are expected to happen for our new default location settings.