Added migration for locations app i18n changes
[ Added by @sevein ]: during the i18n work we did in the Storage Service
we made some changes to our Django models but we forgot to commit the
corresponding database migrations. The missing migration is added in this
commit - however I don't think it actually affects the schema of the database
but it's still a required step because Django needs any model change to be
recorded as a migration. It's important to add missing migrations like this
because we're planning to add a check in CI for this purpose.