Stop looking up DB envs in base settings module
This commit moves database configuration environment lookups to the production
(`production.py`) and development (`local.py`) settings module. Before, the
lookups were in `base.py` which also affected `test.py`. By making this change
we simplify the configuration of the test environment by not requiring to define
database configuration parameters as we prefer to default to SQLite's `:memory`
backend.