Tests: add admin user to fixtures
Storage Service used to automatically add a `test` user to the database. This
behaviour was removed in JiscRDSS/SS for security reasons but it broke our
tests because they rely on that user. In this commit we make the user available
again for the tests by adding it to the fixtures - this won't affect
deployments.